This PR fixes https://github.com/hyrodium/BasicBSpline.jl/issues/313.